pthread library intrinsics (#582)

+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/processing/GEPToAddition.java
@@ -12,7 +12,6 @@ import com.dat3m.dartagnan.program.Program;
 import com.dat3m.dartagnan.program.event.core.utils.RegReader;
 import com.dat3m.dartagnan.program.memory.MemoryObject;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public class GEPToAddition implements ProgramProcessor {
@@ -53,7 +52,7 @@ public class GEPToAddition implements ProgramProcessor {
             assert offsets.size() > 0;
             result = expressions.makeADD(result,
                     expressions.makeMUL(
-                            expressions.makeValue(BigInteger.valueOf(types.getMemorySizeInBytes(type)), archType),
+                            expressions.makeValue(types.getMemorySizeInBytes(type), archType),
                             expressions.makeIntegerCast(offsets.get(0).accept(this), archType, true)));
             for (final Expression oldOffset : offsets.subList(1, offsets.size())) {
                 final Expression offset = oldOffset.accept(this);
@@ -61,7 +60,7 @@ public class GEPToAddition implements ProgramProcessor {
                     type = arrayType.getElementType();
                     result = expressions.makeADD(result,
                             expressions.makeMUL(
-                                    expressions.makeValue(BigInteger.valueOf(types.getMemorySizeInBytes(arrayType.getElementType())), archType),
+                                    expressions.makeValue(types.getMemorySizeInBytes(arrayType.getElementType()), archType),
                                     expressions.makeIntegerCast(offset, archType, true)));
                     continue;
                 }
@@ -78,7 +77,7 @@ public class GEPToAddition implements ProgramProcessor {
                 for (final Type elementType : aggregateType.getDirectFields().subList(0, value)) {
                     o += types.getMemorySizeInBytes(elementType);
                 }
-                result = expressions.makeADD(result, expressions.makeValue(BigInteger.valueOf(o), archType));
+                result = expressions.makeADD(result, expressions.makeValue(o, archType));
             }
             return result;
         }
